feat: implement loadOldStorageKey utility and update context loading methods
